requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-cors==4.0.0
python-dotenv==1.0.0
mysqlclient==2.2.0
bcrypt==4.0.1
gunicorn==21.2.0
flask_bcrypt 
requests 
beautifulsoup4 
pymysql 